Improved ScopeTimers Testing, and restructed folders and files.
